+++ smrt/src/main/java/com/smart/dto/ColisRequestDTO.java
@@ -5,7 +5,6 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -14,7 +13,7 @@ public class ColisRequestDTO {
     private String description;
     
     @NotNull(message = "Le poids est requis")
-    private BigDecimal poids;
+    private Double poids;
     
     private String priorite; // HAUTE, MOYENNE, BASSE
